pyproject.toml
[project]
name = "shadowscribe"
version = "2.0.0"
description = "Comprehensive D&D/RPG character management system with RAG-powered knowledge base"
requires-python = "==3.12.*"
dependencies = [
    "typing-extensions>=4.0.0",
    "dataclasses-json>=0.6.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "setuptools",
    "replicate>=0.25.0",
    "pdf2image>=1.16.3",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "numpy>=1.24.0",
    # Local classifier dependencies
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "scikit-learn>=1.3.0",
    "safetensors>=0.4.0",
    "dacite>=1.9.2",
    "sentence-transformers>=5.1.2",
    "rank-bm25>=0.2.2",
    # Model download from Google Drive
    "gdown>=5.0.0",
    "firebase-admin>=7.1.0",
    "google-cloud-storage>=3.7.0",
    "google-cloud-firestore>=2.19.0",
    "pymysql>=1.1.2",  # Keep for migration script
    "uvicorn>=0.40.0",
    "fastapi>=0.127.0",
    "websockets>=15.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
